scientific computing in Python;
I am ambivalent about Python, and don't have the cycles to learn it
properly.  But I do like scientific computing, and I have the motivation
to get better-versed in that.  So I'd better read-up on NumPy and SciPy,
since the scientific computing community seems to have anointed Python
as their platform of choice.