finally getting back to Montgomery Reduction;
I first discovered Montgomery Reduction a number of years ago when I was
attempting to read through the code of OpenSSL.  There was no
satisfactory explanation on Wikipedia at the time, and the seminal paper
on it was a little bit over my head.  But now it looks like the
Wikipedia article has gotten some needed updates.